# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API tasks",
        "description": "Management tasks",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Signup user",
                "operationId": "aaee79251c237baf5b28cd58be60b12a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "number",
                                        "example": "text"
                                    },
                                    "email": {
                                        "type": "number",
                                        "example": "text"
                                    },
                                    "password": {
                                        "type": "number",
                                        "example": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "token": {
                                                        "type": "string",
                                                        "example": "text"
                                                    },
                                                    "user": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "text"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "example": "text"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2023-02-23T12:33:45.000000Z"
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2023-02-23T12:33:45.000000Z"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        },
                                        "data": {
                                            "type": "list",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login user",
                "operationId": "4a16cbeb901087664661f8f8b3f4c012",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "number",
                                        "example": "text"
                                    },
                                    "password": {
                                        "type": "number",
                                        "example": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        },
                                        "data": {
                                            "type": "list",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "logout user",
                "operationId": "392b7d1309524bd8a5cb5272d91f6a5e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}